Changed the glacier volume inversion method
I changed the target for the volume inversion.  Instead of trying
to match the total glacier volume from the input file, CISM now tries
to match the volume over the observed glacier footprint, i.e., the region
covered by the initial glacier.  This is a new variable in the glacier
derived type, called volume_over_init_region.

The goal is to avoid cancelling errors in the volume inversion.
If a glacier advances, its area will increase, so if CISM is simply
trying to match the total volume, it will tend to make the interior
of the glacier too thin (and conversely for retreating glaciers).
With the new method, CISM will generate glaciers that have the correct
thickness over the observed footprint.